requirements.txt
pygame-ce>=2.5.6
pygame-gui>=0.6.14
numpy>=1.26.4
matplotlib>=3.10.0
tqdm>=4.67.3
Pillow>=10.4.0
pytest>=9.0.2